Solve error which occures when no picture is added when creatin a new auction.
