+++ baseProject/src/main/java/ch/unibe/ese/team1/model/Ad.java
@@ -1,9 +1,29 @@
 package ch.unibe.ese.team1.model;
 
+import java.util.List;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.ManyToMany;
+
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
 
 /** Describes an advertisement that users can place and search for. */
 @Entity
 public class Ad extends Advertisement{
 
+	@Fetch(FetchMode.SELECT)
+	@ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+	private List<AdPicture> pictures;
+	
+	public List<AdPicture> getPictures() {
+		return pictures;
+	}
+
+	public void setPictures(List<AdPicture> pictures) {
+		this.pictures = pictures;
+	}
+	
 }
