+++ baseProject/src/main/java/ch/unibe/ese/team1/model/AuctionPicture.java
@@ -1,0 +1,12 @@
+package ch.unibe.ese.team1.model;
+
+import javax.persistence.Entity;
+
+/**
+ * Represents a picture that is linked to an auction and therefore is linked to an
+ * auction.
+ */
+@Entity
+public class AuctionPicture extends Picture {
+
+}
